Added fixes
Calendar view filtering functionality added.
Fixed overflow issues on listing events on calendar view.
Added recent searches feature.
